requirements.txt
google-generativeai
google-cloud-aiplatform
google-cloud-storage
google-adk
pypdf 
python-dotenv
pypdf
pandas
python-dotenv
matplotlib
seaborn